+++ packages/navi-js/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @awell-health/navi-js
 
+## 0.0.5-beta.2
+
+### Patch Changes
+
+- Update handlers
+- Updated dependencies
+  - @awell-health/navi-core@0.0.5-beta.2
+
 ## 0.0.5-beta.1
 
 ### Patch Changes