+++ packages/navi-react/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @awell-health/navi-js-react
 
+## 0.0.5-beta.2
+
+### Patch Changes
+
+- Update handlers
+- Updated dependencies
+  - @awell-health/navi-core@0.0.5-beta.2
+  - @awell-health/navi-js@0.0.5-beta.2
+
 ## 0.0.5-beta.1
 
 ### Patch Changes